+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 
 ### Fixed
+- `axi_sim_mem`: Keep R response stable while not accepted.
 
 
 ## 0.34.0 - 2022-03-09