+++ CHANGELOG.md
@@ -5,13 +5,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
-## Unreleased
+## 0.35.0 - 2022-03-11
 
 ### Added
 - `axi_sim_mem`: Add monitoring interface to observe the point of coherency between the write and
   the read channel.
-
-### Changed
 
 ### Fixed
 - `axi_sim_mem`: Keep R response stable while not accepted.